update examples to make unittests pass
