+++ doc/index.rst
@@ -41,13 +41,14 @@ Example
     >>> import sklearn.model_selection
     >>> import sklearn.datasets
     >>> import sklearn.metrics
-    >>> X, y = sklearn.datasets.load_digits(return_X_y=True)
-    >>> X_train, X_test, y_train, y_test = \
-            sklearn.model_selection.train_test_split(X, y, random_state=1)
-    >>> automl = autosklearn.classification.AutoSklearnClassifier()
-    >>> automl.fit(X_train, y_train)
-    >>> y_hat = automl.predict(X_test)
-    >>> print("Accuracy score", sklearn.metrics.accuracy_score(y_test, y_hat))
+    >>> if __name__ == "__main__":
+    >>>     X, y = sklearn.datasets.load_digits(return_X_y=True)
+    >>>     X_train, X_test, y_train, y_test = \
+                sklearn.model_selection.train_test_split(X, y, random_state=1)
+    >>>     automl = autosklearn.classification.AutoSklearnClassifier()
+    >>>     automl.fit(X_train, y_train)
+    >>>     y_hat = automl.predict(X_test)
+    >>>     print("Accuracy score", sklearn.metrics.accuracy_score(y_test, y_hat))
 
 
 This will run for one hour and should result in an accuracy above 0.98.
