+++ examples/20_basic/example_regression.py
@@ -13,36 +13,37 @@ import sklearn.metrics
 import autosklearn.regression
 
 
-############################################################################
-# Data Loading
-# ============
+if __name__ == "__main__":
+    ############################################################################
+    # Data Loading
+    # ============
 
-X, y = sklearn.datasets.load_boston(return_X_y=True)
+    X, y = sklearn.datasets.load_boston(return_X_y=True)
 
-X_train, X_test, y_train, y_test = \
-    sklearn.model_selection.train_test_split(X, y, random_state=1)
+    X_train, X_test, y_train, y_test = \
+        sklearn.model_selection.train_test_split(X, y, random_state=1)
 
-############################################################################
-# Build and fit a regressor
-# =========================
+    ############################################################################
+    # Build and fit a regressor
+    # =========================
 
-automl = autosklearn.regression.AutoSklearnRegressor(
-    time_left_for_this_task=120,
-    per_run_time_limit=30,
-    tmp_folder='/tmp/autosklearn_regression_example_tmp',
-    output_folder='/tmp/autosklearn_regression_example_out',
-)
-automl.fit(X_train, y_train, dataset_name='boston')
+    automl = autosklearn.regression.AutoSklearnRegressor(
+        time_left_for_this_task=120,
+        per_run_time_limit=30,
+        tmp_folder='/tmp/autosklearn_regression_example_tmp',
+        output_folder='/tmp/autosklearn_regression_example_out',
+    )
+    automl.fit(X_train, y_train, dataset_name='boston')
 
-############################################################################
-# Print the final ensemble constructed by auto-sklearn
-# ====================================================
+    ############################################################################
+    # Print the final ensemble constructed by auto-sklearn
+    # ====================================================
 
-print(automl.show_models())
+    print(automl.show_models())
 
-###########################################################################
-# Get the Score of the final ensemble
-# ===================================
+    ###########################################################################
+    # Get the Score of the final ensemble
+    # ===================================
 
-predictions = automl.predict(X_test)
-print("R2 score:", sklearn.metrics.r2_score(y_test, predictions))
+    predictions = automl.predict(X_test)
+    print("R2 score:", sklearn.metrics.r2_score(y_test, predictions))
